requirements.txt
asn1crypto==0.24.0
asyncmongo==1.3
attrs==18.1.0
Automat==0.6.0
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
deskapi==0.1
Django==2.0.5
entrypoints==0.2.3
Flask==1.0.2
gevent==1.3.0
greenlet==0.4.13
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
itsdangerous==0.24
jedi==0.12.0
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
linecache2==1.0.0
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.3
pandas==0.23.0
pandoc==1.0.2
pandocfilters==1.4.2
parsel==1.4.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
ply==3.11
progressbar2==3.37.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.2.0
pymongo==3.6.1
pyOpenSSL==17.5.0
pypandoc==1.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
scipy==1.1.0
Scrapy==1.5.0
scrapy-splash==0.7.2
Send2Trash==1.5.0
service-identity==17.0.0
shuttle==0.1
simplegeneric==0.8.1
six==1.11.0
snownlp==0.12.3
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traceback2==1.4.0
traitlets==4.3.2
Twisted==18.4.0
txlib==0.0.0
unittest2==1.1.0
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
zope.interface==4.5.0
